[Silver V] Title: 귀찮음, Time: 48 ms, Memory: 5816 KB, Score: 30 point -BaekjoonHub

+++ 백준/Silver/16208. 귀찮음/귀찮음.cc
@@ -15,8 +15,8 @@ int main()
 
 	int n;
 	cin >> n;
-	vector<int> v(n);
-	int sum = 0;
+	vector<ll> v(n);
+	ll sum = 0;
 	for (int i = 0; i < n; i++)
 	{
 		cin >> v[i];
@@ -25,7 +25,7 @@ int main()
 
 	sort(v.begin(), v.end());
 
-	int ans = 0;
+	ll ans = 0;
 	for (int i = 0; i < n - 1; i++)
 	{
 		sum -= v[i];
